Migrate template pages
