+++ packages/frontend/pages/profile/transaction-templates/add.tsx
@@ -1,0 +1,7 @@
+import { AddTemplateContainer } from '$container/templates/add-template.container';
+
+const AddTemplatePage = () => {
+  return <AddTemplateContainer />;
+};
+
+export default AddTemplatePage;
