+++ packages/frontend/pages/profile/transaction-templates/index.tsx
@@ -1,0 +1,7 @@
+import { TemplateListingContainer } from '$container/templates/template-listing.container';
+
+const TemplateListingPage = () => {
+  return <TemplateListingContainer />;
+};
+
+export default TemplateListingPage;
